Add the ".svn/" pattern to the .gitignore file
It happened to me after an svn project copy (ie.: copy of the whole plugin
folder) to face this situation, thus it is better to eliminate without any
drawback.

git.kde.org defends against ".svn/" committing in fact, but it is still a good
defense on our own.
https://projects.kde.org/projects/repo-management/repository/revisions/master/entry/config/blockedfiles.cfg

If the need arises, we can add later ".svn*" or something like that, but as for
the time being, ".svn/" is the most common pattern that can cause a trouble with
the aforementioned scenario.